// src/app/shared/models/scenario.model.ts
export enum Scenario {
    EARLY_REPUBLIC = 'Early Republic',
    MIDDLE_REPUBLIC = 'Middle Republic',
    LATE_REPUBLIC = 'Late Republic',
    EARLY_TO_LATE_REPUBLIC = 'Early to Late Republic',
    EARLY_TO_MIDDLE_REPUBLIC = 'Early to Middle Republic',
    MIDDLE_TO_LATE_REPUBLIC = 'Middle to Late Republic',
    ALTERNATE_HISTORY = 'Alternate History'
}